Adding objects
The TextureLoader is now static and return a Texture object wich is more judicious. The GamerLogger is now enhanced, all the standard errors are redirected to it. I have added a Tile object wich can be easely placed on screen by just putting the x and y coordinates, a Texture and a size. Updated the .gitignore