Added Shaders
I have added a ShaderLoader which is used to load the Vertex and Fragment shaders located in only one file and then compile them to prepare them for use. This ShaderLoader class return a Shader object wich can be used with the Tile object. For now the default shader can' be used for textures but only for colors.