A lot of changes
The tile system has been completly redone (and is not done yet). The GLFWWindow is no more initialized with the glOrtho() because it was causing problems with shaders. To repair that I made the Camera object which can be configured to view a certain amount of tiles. There is new functions into the Time object for getting the deltaTime calculated in the loop and set it in the class itself. For now the Tile class is doing a default tile on the screen because I am currently checking if the system works properly. The VBO class has been completely redone to draw and use the default shader correctly.