Tile changes
The Tile object can now be used with a Color object that juste contains float values between 0.0f and 1.0f to change the color of the drawn Tile. I have added some comments to the Camera to explain how it works. I have added some static constant for the window modes.