Entity Component System implementation and SceneManager
I have implemented an ETS wich seems to be a standard in the video games industry. This is the one I like the most. Now, you can create a new scene and register it to the SceneManager. In each scene you can create some game objects that you can fill with components. One of these component (and the only one for now) is the Tile object wich now extend the Component class. Like that each Component that are instanciated contains there own loop wich is connected to the main loop. I have removed all the Buffers list because the Garbage Collector will automatically delete theme when there are not used.