Batch rendering
I have implemented a way to do batch rendering wich can improve a lot performances when drawing multiple Surface.
I have added a method to get a List of multiple components of the same type into the GameObject class.
Implemented the Renderer because in the future I will implement a zIndex for the BatchRenderer.
The ResourceManager has been modified : Before the handlers were a String, now it is the class of the Object you want to work with.
I have added a primitive Component to the engine wich is calles the Surface for drawing quad shapped element on the screen. This surface is very useful because it use the BatchRenderer.