BatchRenderer and dirty flag
Now, the BatchRenderer will update his buffer only if a surface changed there properties. It is what's called a dirty flag. The SpritesheetLoader was not working because when extracting values from a JSON file it returns long ans not int. Some equals methods has been added to Texture, Color and Transform. You can now create a Surface with custom texture coordinates (for exemple using a Spritesheet).
PS : Mario is a little drunk