Font rendering
I have implemented a way to draw text on the screen BUT I am not confident with the way I have done that because the Text object is static and I think it is strange but at least there is now a base of this implementation.
I have added a couple .ttf files to test the thing.
Until now all the textures were loaded upside down (YES I KNOW THIS IS COMPLETELY STUPID), NOW it is fixed. Same thing with the spritesheets.