Text rendering
The Text rendering and not Font rendering xD has been modified because it was not well made before. Now the Text object is an instanciable object and can be used by putting it in a surface. Then you can change the text, change the color etc... I have updated the Time static object because there is a glfwGetTime() function available wich is far better that what I had done.
The WIDTH and HEIGHT are now static in the GLFWWindow.