LineBatchRenderer
I have added a LineBatchRenderer and renamed the BatchRenderer.java to SurfaceBatchRenderer.java. I was not planning to do multiple BatchRenderer but I think I will do that for every primitive elements because they have to perform very fast. I prefer keep the computer performance for the user of this engine. I had to create a special shader for this part because the default shader can only be used with textureID and in this case there is no textures IDs. I have added some class files just to know what to do. These are mostly GUI elements. I renamed the Text.java to Label.java since it is more appropriate.