Bug fixes
In the SurfaceBatchRenderer I was using == operator and not equals. The == operator verify for pointer equality and not for 'content' equality.
Some methods have been moved from Label to TextureLoader since they are useful for other things.
The Texture object now require a String containing the filePath and a ByteBuffer containing the entire image in bytes.