Event System
I have implement an Event System. This System is not in the same Thread. Actually I am not sure of the EventSystemI interface name, it seems a bit unsuitable. I am working on the Button implementation and it almost work but, the mouse coordinate system is not the same as the renderers coordinates so it is a little bit strange.