Big changes on the Label System
The Label System has been changed to a more optimized one. Now we use Font Atlas wich is basically a texture that contains all characters of a Font the to pick a character, we use the Texture Coordinates. The label drawing system is to review because it is now very efficient and very restrictive.