Label enhancement WIP
Last time I forgot to make a proper commit. So I have create a TextRenderer specially for the on-screen text rendering because I was not very confident with way I had done it because I was basically making surfaces and rendering all of them etc... This was not efficient and not practical. Now the new system is far better because now I just create 'custom' vertices for the texts and that's far better.