TextRendering
After a lot of thinking I have finally made a system that works and almost fully optimized. The only thing is that for each Label a new TextRenderer is made. Maybe I can optimize it by doing a batcher instead ? For now it is very good I think. No more ******* memory leaks.
I have added a new ButtonEvent which is fired when the user 'unhover' a button.
I have added a bunch of default Colors in the Engine class.