Rendering changes and improvement
The shaders are now upgraded to GLSL 3.3 core because, this engine is compatible with OpenGL 3.0 and above, there is no point to downgrade de shaders.
I am starting to work on FrameBuffers.
You can draw a Label with a custom ZIndex
The RendererHelper has been almost completely redone because I have added a Renderers interface to support 'global' Z Indexing for all components.